+++ api/models/SupplierShipOrder.js
@@ -263,6 +263,11 @@ module.exports = {
       allowNull: false,
     },
 
+    shippingTelephone: {
+      type: Sequelize.STRING(32),
+      allowNull: false,
+    },
+
 		createdDateTime: {
 			type: Sequelize.VIRTUAL,
 			get: function() {
@@ -315,12 +320,11 @@ module.exports = {
       type: Sequelize.VIRTUAL,
       get: function(){
         try{
-          let total = this.getDataValue('total');
-          if(!total){
+          let tax = this.getDataValue('tax');
+          if(!tax){
             return '';
           }
-          total = Math.round(total * 0.05);
-          return UtilsService.moneyFormat(total);
+          return UtilsService.moneyFormat(tax);
 
         } catch(e){
           sails.log.error(e);
@@ -331,12 +335,11 @@ module.exports = {
       type: Sequelize.VIRTUAL,
       get: function(){
         try{
-          let total = this.getDataValue('total');
-          if(!total){
+          let totalIncludeTax = this.getDataValue('totalIncludeTax');
+          if(!totalIncludeTax){
             return '';
           }
-          total = Math.round(total * 1.05);
-          return UtilsService.moneyFormat(total);
+          return UtilsService.moneyFormat(totalIncludeTax);
 
         } catch(e){
           sails.log.error(e);